pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ttrecon"
version = "0.0.5.post2"
description = "TT-RECON: Truth-first targeted therapy recon pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=1.10",
  "PyYAML>=6.0",
  "Jinja2>=3.1",
  "requests>=2.31",
]

[project.scripts]
ttrecon = "ttrecon.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
